Suppress Compiler Warnings - Compare of mixed signedness
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@6957 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: def9a411943451e5b2f6555e8779eb88327ae870
Former-commit-id: 4909b6226d2f95aba29ed8d4ed783020034daa68

+++ src/drivers/shadow/src/Driver.cpp
@@ -383,7 +383,7 @@ void	Driver::InitTrack(
 
             LogSHADOW.debug( "%s: %g, %g\n", buf, innerModA, innerModB );
 
-            while( m_priv[p].INNER_MOD.size() < bend * 2 + 2 )
+            while (m_priv[p].INNER_MOD.size() < static_cast<size_t>(bend * 2 + 2))
                 m_priv[p].INNER_MOD.push_back( 0 );
 
             if( innerModA != 999 )
